nfsd: Turn off post-op attributes for READ when re-exporting NFS
If NFS is being re-exported, then it would be nice to return post-op
attributes, but the problem is we cannot return them consistently
without impacting performance. This causes problems for some NFSv3
clients, since they are not able to reliably determine the header size
in order to place the read payload correctly. In some cases, it actually
leads to EIO errors.
For that reason, let's just turn off post-op attributes for READ when
re-exporting NFS (and keep it in all other cases).

Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>
Signed-off-by: Mike Snitzer <snitzer@hammerspace.com>