NFSv4/pnfs: Fix setattr behaviour when pNFS is enabled
If pNFS flexfiles is in use, and the user tries to explicitly set a
timestamp, we want to return either just the read/write layout or all
layouts depending on if we're setting the mtime or the atime.

Since the blocks driver also has pNFS-specific layoutreturn behaviour in
setattr that differs from that of flexfiles, we should just add a
callback in which the two drivers can do the right thing without further
polluting the generic NFSv4 code.

Fixes: d67ae825a59d ("pnfs/flexfiles: Add the FlexFile Layout Driver")
Signed-off-by: Trond Myklebust <trond.myklebust@hammerspace.com>