nfs: move definition of NFS_DELEGATION_FLAG_TIME within nfs/delegation.h
This fixes a rebase porting issue where NFS_DELEGATION_FLAG_TIME was
left defined within the #if IS_ENABLED(CONFIG_NFS_V4) block of
delegation.h rather than outside it.  This change really doesn't
matter or change anything given we obviously enable CONFIG_NFS_V4, but
best to fix this up so the code is correct (and matches code that was
previously in mainline).

Fixes: 567714deb78ac ("NFSv4: Add support for delegated atime and mtime attributes")
Signed-off-by: Mike Snitzer <snitzer@hammerspace.com>